Program now applies a guassian blur before the sobel op, and also removes alpha=0 pixels
Since the sobel operator is somewhat sensitive to noise, I added in a gaussian convolution to remove higher frequencies. This drastically improves the clarity of the final image, although this is still dependent on the configuration of the gaussian kernel.

Additionally, I've added to the grayscale portion so that the pixel RGB values are now turned off if the alpha value is 0. This removes any invisible edges hidden within the image.